# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Gateway Registry",
    "description": "A registry and management system for Model Context Protocol (MCP) servers",
    "version": "1.0.11"
  },
  "paths": {
...
    "/api/servers": {
      "get": {
        "tags": ["Server Management"],
        "summary": "Get Servers Json",
        "description": "Get servers data as JSON for React frontend and external API (supports both session cookies and Bearer tokens).",
        "operationId": "get_servers_json_api_servers_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/servers/groups": {
      "get": {
        "tags": ["Server Management"],
        "summary": "List Groups Api",
        "description": "List all scope groups via JWT Bearer Token authentication (External API).\n\nThis endpoint provides the same functionality as GET /api/internal/list-groups\nbut uses modern JWT Bearer token authentication.\n\n**Authentication:** JWT Bearer token (via nginx X-User header)\n**Authorization:** Requires valid JWT token from auth system\n\n**Response:**\nReturns a list of all groups and their synchronization status.\n\n**Example:**\n```bash\ncurl -X GET https://registry.example.com/api/servers/groups \\\n  -H \"Authorization: Bearer $JWT_TOKEN\"\n```",
        "operationId": "list_groups_api_api_servers_groups_get",
        "parameters": [
          {
            "name": "include_keycloak",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Keycloak"
            }
          },
          {
            "name": "include_scopes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Scopes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/agents": {
      "get": {
        "tags": ["Agent Management"],
        "summary": "List Agents",
        "description": "List all agents filtered by user permissions.\n\nArgs:\n    query: Optional search query\n    enabled_only: Only return enabled agents\n    visibility: Filter by visibility level\n    user_context: Authenticated user context\n\nReturns:\n    List of agent info objects",
        "operationId": "list_agents_api_agents_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search query string",
              "title": "Query"
            },
            "description": "Search query string"
          },
          {
            "name": "enabled_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show only enabled agents",
              "default": false,
              "title": "Enabled Only"
            },
            "description": "Show only enabled agents"
          },
          {
            "name": "visibility",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by visibility",
              "title": "Visibility"
            },
            "description": "Filter by visibility"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/search/semantic": {
      "post": {
        "tags": ["Semantic Search"],
        "summary": "Unified semantic search for MCP servers and tools",
        "description": "Run a semantic search against MCP servers (and their tools) using FAISS embeddings.",
        "operationId": "semantic_search_api_search_semantic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/v0.1/servers": {
      "get": {
        "tags": ["Anthropic Registry API", "Anthropic Registry API"],
        "summary": "List MCP servers",
        "description": "Returns a paginated list of all registered MCP servers that the authenticated user can access.",
        "operationId": "list_servers_v0_1_servers_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pagination cursor",
              "title": "Cursor"
            },
            "description": "Pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum number of items",
              "title": "Limit"
            },
            "description": "Maximum number of items"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentSearchResult": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "trust_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Level"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": false
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "match_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Context"
          }
        },
        "type": "object",
        "required": ["path", "agent_name", "relevance_score"],
        "title": "AgentSearchResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchingToolResult": {
        "properties": {
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score",
            "default": 0.0
          },
          "match_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Context"
          }
        },
        "type": "object",
        "required": ["tool_name"],
        "title": "MatchingToolResult",
        "description": "Tool matching result - basic info for display.\n\nNote: inputSchema is NOT included here to avoid duplication.\nFull tool details including inputSchema are in the tools[] array."
      },
...
      "PaginationMetadata": {
        "properties": {
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor",
            "description": "Cursor for next page"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count",
            "description": "Number of items in current page"
          }
        },
        "type": "object",
        "title": "PaginationMetadata",
        "description": "Pagination information for server lists."
      },
...
      "Repository": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Repository URL for browsing source code"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Repository hosting service identifier (e.g., 'github')"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id",
            "description": "Repository ID from hosting service"
          },
          "subfolder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subfolder",
            "description": "Path within monorepo"
          }
        },
        "type": "object",
        "required": ["url", "source"],
        "title": "Repository",
        "description": "Repository metadata for MCP server source code."
      },
      "SemanticSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Query",
            "description": "Natural language query"
          },
          "entity_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": ["mcp_server", "tool", "a2a_agent"]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Types",
            "description": "Optional entity filters"
          },
          "max_results": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum results per entity collection",
            "default": 10
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "SemanticSearchRequest"
      },
      "SemanticSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerSearchResult"
            },
            "type": "array",
            "title": "Servers"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolSearchResult"
            },
            "type": "array",
            "title": "Tools"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentSearchResult"
            },
            "type": "array",
            "title": "Agents"
          },
          "total_servers": {
            "type": "integer",
            "title": "Total Servers",
            "default": 0
          },
          "total_tools": {
            "type": "integer",
            "title": "Total Tools",
            "default": 0
          },
          "total_agents": {
            "type": "integer",
            "title": "Total Agents",
            "default": 0
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "SemanticSearchResponse"
      },
      "ServerDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Server name in reverse-DNS format"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Server description"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Server version"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Human-readable server name"
          },
          "repository": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Repository"
              },
              {
                "type": "null"
              }
            ],
            "description": "Repository information"
          },
          "websiteUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Websiteurl",
            "description": "Server website URL"
          },
          "packages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Package"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packages",
            "description": "Package distributions"
          },
          "_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta",
            "description": "Extensible metadata"
          }
        },
        "type": "object",
        "required": ["name", "description", "version"],
        "title": "ServerDetail",
        "description": "Detailed MCP server information."
      },
      "ServerList": {
        "properties": {
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerResponse"
            },
            "type": "array",
            "title": "Servers",
            "description": "List of servers"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pagination info"
          }
        },
        "type": "object",
        "required": ["servers"],
        "title": "ServerList",
        "description": "Response for server list queries."
      },
      "ServerResponse": {
        "properties": {
          "server": {
            "$ref": "#/components/schemas/ServerDetail",
            "description": "Server details"
          },
          "_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta",
            "description": "Registry-managed metadata"
          }
        },
        "type": "object",
        "required": ["server"],
        "title": "ServerResponse",
        "description": "Response for single server query."
      },
      "ServerSearchResult": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "server_name": {
            "type": "string",
            "title": "Server Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "num_tools": {
            "type": "integer",
            "title": "Num Tools",
            "default": 0
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": false
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "match_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Context"
          },
          "matching_tools": {
            "items": {
              "$ref": "#/components/schemas/MatchingToolResult"
            },
            "type": "array",
            "title": "Matching Tools"
          }
        },
        "type": "object",
        "required": ["path", "server_name", "relevance_score"],
        "title": "ServerSearchResult"
      },
      "ToolSearchResult": {
        "properties": {
          "server_path": {
            "type": "string",
            "title": "Server Path"
          },
          "server_name": {
            "type": "string",
            "title": "Server Name"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "inputSchema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputschema",
            "description": "JSON Schema for tool input"
          },
          "relevance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relevance Score"
          },
          "match_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Context"
          }
        },
        "type": "object",
        "required": [
          "server_path",
          "server_name",
          "tool_name",
          "relevance_score"
        ],
        "title": "ToolSearchResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token obtained from Keycloak OAuth2 authentication. Include in Authorization header as: `Authorization: Bearer <token>`"
      }
    }
  }
}